+++ casper/src/test/scala/coop/rchain/casper/helper/BlockDataContract.scala
@@ -5,13 +5,13 @@ import coop.rchain.crypto.PublicKey
 import coop.rchain.metrics.Span
 import coop.rchain.models.{ListParWithRandom, Par}
 import coop.rchain.rholang.interpreter.{ContractCall, RhoType}
-import coop.rchain.rholang.interpreter.Runtime.SystemProcess
+import coop.rchain.rholang.interpreter.SystemProcesses.ProcessContext
 
 object BlockDataContract {
   import cats.implicits._
 
   def set[F[_]: Concurrent: Span](
-      ctx: SystemProcess.Context[F]
+      ctx: ProcessContext[F]
   )(message: Seq[ListParWithRandom]): F[Unit] = {
 
     val isContractCall = new ContractCall(ctx.space, ctx.dispatcher)
