+++ casper/src/test/scala/coop/rchain/casper/helper/CasperInvalidBlocksContract.scala
@@ -1,17 +1,16 @@
 package coop.rchain.casper.helper
 
 import cats.effect.Concurrent
-import coop.rchain.crypto.PublicKey
 import coop.rchain.metrics.Span
 import coop.rchain.models.{ListParWithRandom, Par}
-import coop.rchain.rholang.interpreter.{ContractCall, RhoType}
-import coop.rchain.rholang.interpreter.Runtime.SystemProcess
+import coop.rchain.rholang.interpreter.{ContractCall}
+import coop.rchain.rholang.interpreter.SystemProcesses.ProcessContext
 
 object CasperInvalidBlocksContract {
   import cats.implicits._
 
   def set[F[_]: Concurrent: Span](
-      ctx: SystemProcess.Context[F]
+      ctx: ProcessContext[F]
   )(message: Seq[ListParWithRandom]): F[Unit] = {
 
     val isContractCall = new ContractCall(ctx.space, ctx.dispatcher)
