+++ casper/src/test/scala/coop/rchain/casper/helper/Secp256k1SignContract.scala
@@ -4,13 +4,13 @@ import cats.effect.Concurrent
 import coop.rchain.crypto.signatures.Secp256k1
 import coop.rchain.metrics.Span
 import coop.rchain.models.ListParWithRandom
-import coop.rchain.rholang.interpreter.Runtime.SystemProcess
+import coop.rchain.rholang.interpreter.SystemProcesses
 import coop.rchain.rholang.interpreter.{ContractCall, RhoType}
 
 object Secp256k1SignContract {
 
   def get[F[_]: Concurrent: Span](
-      ctx: SystemProcess.Context[F]
+      ctx: SystemProcesses.ProcessContext[F]
   )(message: Seq[ListParWithRandom]): F[Unit] = {
     val isContractCall = new ContractCall(ctx.space, ctx.dispatcher)
     message match {
