+++ casper/src/test/scala/coop/rchain/casper/helper/SysAuthTokenContract.scala
@@ -4,7 +4,7 @@ import cats.effect.Concurrent
 import coop.rchain.metrics.Span
 import coop.rchain.models.{GSysAuthToken, ListParWithRandom}
 import coop.rchain.rholang.interpreter.{ContractCall, RhoType}
-import coop.rchain.rholang.interpreter.Runtime.SystemProcess
+import coop.rchain.rholang.interpreter.SystemProcesses.ProcessContext
 
 /**
   * Warning: This should under no circumstances be available in production
@@ -13,7 +13,7 @@ object SysAuthTokenContract {
   import cats.implicits._
 
   def get[F[_]: Concurrent: Span](
-      ctx: SystemProcess.Context[F]
+      ctx: ProcessContext[F]
   )(message: Seq[ListParWithRandom]): F[Unit] = {
 
     val isContractCall = new ContractCall(ctx.space, ctx.dispatcher)
