+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/accounting/HasCost.scala
@@ -1,0 +1,5 @@
+package coop.rchain.rholang.interpreter.accounting
+
+trait HasCost[F[_]] {
+  def cost: _cost[F]
+}
